requirements.txt
openai-whisper
moviepy
pydub
python-docx
scikit-learn
pyinstaller
pdfplumber